@@ -9,7 +9,6 @@ import (
 	"time"
 
 	cstructs "github.com/hashicorp/nomad/client/structs"
-	"github.com/hashicorp/nomad/helper"
 	"github.com/hashicorp/nomad/nomad/structs"
 	"github.com/hashicorp/nomad/plugins/drivers"
 )
@@ -26,7 +25,7 @@ func NewDriverHandle(
 		net:         net,
 		taskID:      taskID,
 		killSignal:  task.KillSignal,
-		killTimeout: helper.Min(task.KillTimeout, maxKillTimeout),
+		killTimeout: min(task.KillTimeout, maxKillTimeout),
 	}
 }
 
