@@ -500,6 +500,9 @@ func TestNetworkFingerPrint_MultipleAliases(t *testing.T) {
 	for alias := range cfg.HostNetworks {
 		expected = append(expected, alias)
 	}
+	// eth3 matches the NetworkInterface and will then generate the 'default'
+	// alias
+	expected = append(expected, "default")
 	sort.Strings(expected)
 	sort.Strings(aliases)
 	require.Equal(t, expected, aliases, "host networks should match aliases")
@@ -537,7 +540,7 @@ func TestNetworkFingerPrint_HostNetworkReservedPorts(t *testing.T) {
 					CIDR:      "100.64.0.11/10",
 				},
 			},
-			expected: []string{"", "", ""},
+			expected: []string{"", "", "", ""},
 		},
 		{
 			name: "reserved ports in some aliases",
@@ -560,7 +563,7 @@ func TestNetworkFingerPrint_HostNetworkReservedPorts(t *testing.T) {
 					CIDR:      "100.64.0.11/10",
 				},
 			},
-			expected: []string{"22", "80,3000-4000", ""},
+			expected: []string{"22", "80,3000-4000", "", ""},
 		},
 	}
 
