@@ -187,12 +187,11 @@ func (c *VolumeStatusCommand) csiFormatVolumes(vols []*api.CSIVolumeListStub) (s
 // Format the volumes, assumes that we're already sorted by volume ID
 func csiFormatSortedVolumes(vols []*api.CSIVolumeListStub) (string, error) {
 	rows := make([]string, len(vols)+1)
-	rows[0] = "ID|Name|Namespace|Plugin ID|Schedulable|Access Mode"
+	rows[0] = "ID|Name|Plugin ID|Schedulable|Access Mode"
 	for i, v := range vols {
-		rows[i+1] = fmt.Sprintf("%s|%s|%s|%s|%t|%s",
+		rows[i+1] = fmt.Sprintf("%s|%s|%s|%t|%s",
 			v.ID,
 			v.Name,
-			v.Namespace,
 			v.PluginID,
 			v.Schedulable,
 			v.AccessMode,
