@@ -293,7 +293,7 @@ func latestEvalIndex(eval *structs.Evaluation) uint64 {
 		return 0
 	}
 
-	return helper.Max(eval.CreateIndex, eval.SnapshotIndex)
+	return max(eval.CreateIndex, eval.SnapshotIndex)
 }
 
 // missedUnblock returns whether an evaluation missed an unblock while it was in
@@ -545,9 +545,9 @@ func (b *BlockedEvals) unblock(computedClass, quota string, index uint64) {
 
 	// Every eval that has escaped computed node class has to be unblocked
 	// because any node could potentially be feasible.
-	numEscaped := len(b.escaped)
 	numQuotaLimit := 0
-	unblocked := make(map[*structs.Evaluation]string, helper.Max(numEscaped, 4))
+	numEscaped := len(b.escaped)
+	unblocked := make(map[*structs.Evaluation]string, max(uint64(numEscaped), 4))
 
 	if numEscaped != 0 && computedClass != "" {
 		for id, wrapped := range b.escaped {
