@@ -10,8 +10,6 @@ import (
 	log "github.com/hashicorp/go-hclog"
 	memdb "github.com/hashicorp/go-memdb"
 	"github.com/hashicorp/go-set"
-
-	"github.com/hashicorp/nomad/helper"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
 
@@ -263,7 +261,7 @@ func (p *propertySet) UsedCount(option *structs.Node, _ string) (string, string,
 // existing and proposed allocations. It also takes into account any stopped
 // allocations
 func (p *propertySet) GetCombinedUseMap() map[string]uint64 {
-	combinedUse := make(map[string]uint64, helper.Max(len(p.existingValues), len(p.proposedValues)))
+	combinedUse := make(map[string]uint64, max(len(p.existingValues), len(p.proposedValues)))
 	for _, usedValues := range []map[string]uint64{p.existingValues, p.proposedValues} {
 		for propertyValue, usedCount := range usedValues {
 			targetPropertyValue := p.targetedPropertyValue(propertyValue)
